+++ lib/server/controllers/index.js
@@ -23,7 +23,7 @@ const RunsController = require('./runs.controller');
 const StatusController = require('./status.controller');
 const SubsystemsController = require('./subsystems.controller');
 const TagsController = require('./tags.controller');
-const LhcPeriodsController = require('./period.controller');
+const LhcPeriodsController = require('./lhcPeriod.controller');
 
 module.exports = {
     AttachmentsController,
