+++ lib/server/controllers/lhcPeriod.controller.js
@@ -11,7 +11,7 @@
  * or submit itself to any jurisdiction.
  */
 
-const { lhcPeriodService } = require('../../server/services/lhcPeriod/LhcPeriodService.js');
+const { lhcPeriodService } = require('../services/lhcPeriod/LhcPeriodService.js');
 const { adaptFindAndCountAllInService } = require('../utilities/adaptFindAndCountAllInService.js');
 
 /**
@@ -22,9 +22,10 @@ const { adaptFindAndCountAllInService } = require('../utilities/adaptFindAndCoun
  * @returns {undefined}
  */
 const listPeriodsHandler = async (req, res, next) => {
+    // TODO
     // const validatedDTO = await validateDtoOrRepondOnFailure(stdDataRequestDTO, req, res);
     // if (validatedDTO) {
-    res.json(adaptFindAndCountAllInService(await lhcPeriodService.getAllForPhysicsRuns(req.query)));
+    res.json(adaptFindAndCountAllInService(await lhcPeriodService.getAllForPhysicsRuns(req)));
     // }
 };
