+++ lib/server/services/lhcPeriod/LhcPeriodService.js
@@ -49,8 +49,11 @@ class LhcPeriodService {
         queryBuilder.limit(limit);
         queryBuilder.offset(offset);
 
+        console.log('TOBECERTAIN', query)
+        console.log('TOBECERTAIN', filter)
         if (filter) {
             const { ids, names } = filter;
+            console.log(ids)
             if (ids) {
                 queryBuilder.where('id').oneOf(ids.split(ApiConfig.searchItemsSeparator));
             }
